Do not use observer to await strands (fixes #98).
It turns out this is relatively easy to fix by implementing this directly in the
strands await implementation.

`Strand` already extended `AwaitableProvider`. `StrandTrait` now provides
an implementation of `Awaitable` as well, and `awaitable()` simply return
`$this`.

I've added tests to ensure that if an object implements both `Awaitable` and
`AwaitableProvider` the `await()` method is always favoured. This removes the
unecessary overhead of calling `awaitable()` in the case of `StrandTrait`
without requiring that future strand implementations implement `Awaitable`
directly.